Updated some routes to make them more sensible
Updated some URLs to make them call proper things.

Changed the DB for the rooms so that they're stored in tempdb

Changed the DB for the members so they're in TempDB

Updated the rooms so that it links to the chat.

Will edit the room creation slightly for next iteration.

Waiting for Dan to change users over to the other DB

Need to update the AJAX links from chat so that they work.