Fixed our broken AJAX routes the quick way: By changing them all abck aside from the access route itself. We'll take a look at this when time permits but it functions and that's the main thing
Commented out a width entry in the CSS for a button in chat.css because it was cutting off the text. Made note of it in comments of chat.css

Fixed some more whitespace / tab issues.

Fixed a few links that were relative (e.g. stylesheets/imgs/blahblah rather than /stylesheets/imgs/etc.)

All's in working order. Will take a look over the sims now to make sure all the algorithms there are correct with the new users configuration.