Deleted some redundant files, changed room model
Changed the sc-admin view so that it displayed the list of
admins as a drop down populated by actual users rather than just
names.

Changed the sc-admin router so that it added a proper admin
object to the room object. Also changed the room object to take
a proper list of users, e.g. an array of objects. Made a few
minor changes to how the view is rendered, consiquently. I hope
this is okay