Switched motions from being spliced out of the motions array to just being flagged as deleted for reporting purposes
Changed the view accordingly for the chair.

Created a view for motion metrics.

Modified other views accordingly

Created a route that displays motion metrics such as:
- Total motions
- Number rejected
- Number deleted
- Number approved

Found a bug where the chair view will allow a motion
to be both inVote and inDebate at the same time. It seems
these flags are not being set correctly