Changed the comment route to circumvent an early design choice that came from inexperience.
In short, express's pattern matching for routes picks the first
function that could possibly meet the requirement for the function.

In other words, if we have a route that is

POST /debate/:cid/:chid debate.comment

and

POST /debate/vote/:mid debate.vote

the application will always resolve the comment route first even
when trying to call the vote route. This is because we have no
mechanic to differentiate between what is a :cid and what is the
word "vote".

There is an express library called Express Params that can help
with this but we're not using it so our routes have to look like
spaghetti.

OH WELL! It works.