I think metrics are finished.
Right now if a moderator enters the metrics view, they’re given the
option to view metrics by team, general metrics, motion metrics, and
user metrics.
- The user metrics page shows some shit about individual comments, their
  real name, what team they’re on etc.
- The team metrics page is the meat and spuds: This is where you see
  things like total comments for the team, motions moved, number
approved, number denied, number deleted etc. As an example, if argentina
has moved 4, 1 denied and 3 deleted, but australia has moved 1 and 1
approved, they look a little better, right?
- Motions Metrics shows total moved, total denied, total approved, etc.
  and whether or not one is in vote or in debate right now.
- The overview page shows the simulation name, number of participants,
  number of comments in total, number of motions in total for the whole
simulation.

So they kind of drill down as you go through them.

The only thing is that due to our model structure, you can’t get motions
per PERSON. That can’t be tracked and it might be nice to note that as
something we’d add going forward if we wanted. I guess we could add a
user property to motions but ‘mover’ at this point is just the team
name.